+++ packages/insight-viewer/src/utils/cornerstoneHelper/index.ts
@@ -4,5 +4,6 @@ export {
   resize,
   displayImage,
   loadImage,
-  getDefaultViewportForImage
+  getDefaultViewportForImage,
+  getCornerstone
 } from './utils'
